Fix tests for node < 12.10
